Extract jsonp to util function
